some restructuring
