+++ packages/Dockerfile.playwright
@@ -1,4 +1,4 @@
-FROM mcr.microsoft.com/playwright:v1.53.0-jammy
+FROM mcr.microsoft.com/playwright:v1.54.0-jammy
 
 WORKDIR /app/api
