--- packages/api/src/cache/infrastructure/utils/key-transformation.ts
@@ -1,3 +1,0 @@
-export function transformKey(prefix: string, key: string) {
-  return `${prefix}${key.toLowerCase()}`;
-}
