+++ packages/api/src/city/infrastructure/wrappers/city-validation-cache.proxy.ts
@@ -1,7 +1,7 @@
 import { CacheAccessor } from 'src/cache/application/interfaces/cache-service.interface';
-import { transformKey } from 'src/cache/infrastructure/utils/key-transformation';
+import { createCacheKey } from 'src/common/cache/utils/create-cache-key';
 import { HOUR } from 'src/common/utils/time-units';
-import { CityCachePrefixes } from '../../application/constants/weather-cache-prefixes.enum';
+import { CityCachePrefixes } from '../../../common/cache/constants/city-cache-prefixes.enum';
 import { ChainableCityProvider } from '../../application/interfaces/chainable-city.provider';
 
 export class CityProviderCacheProxy extends ChainableCityProvider {
@@ -13,7 +13,7 @@ export class CityProviderCacheProxy extends ChainableCityProvider {
   }
 
   async validateCity(city: string): Promise<boolean> {
-    const cacheKey = transformKey(CityCachePrefixes.CITY_VALIDATION, city);
+    const cacheKey = createCacheKey(CityCachePrefixes.CITY_VALIDATION, city);
     const cached = await this.cacheService.get<boolean>(cacheKey);
     if (cached) return cached; // cache hit
