+++ packages/api/src/common/cache/utils/create-cache-key.ts
@@ -1,0 +1,3 @@
+export function createCacheKey(prefix: string, key: string) {
+  return `${prefix}${key.toLowerCase()}`;
+}
