+++ packages/api/src/common/weather-provider/types/openweather-weather-fetch.type.ts
