+++ packages/api/src/common/weather-provider/types/weatherapi-city-fetch.type.ts
