+++ packages/api/src/weather/infrastructure/providers/weatherapi.provider.ts
@@ -5,7 +5,7 @@ import { AxiosError } from 'axios';
 import { firstValueFrom } from 'rxjs';
 import { CityNotFoundException } from 'src/common/errors/city-not-found.error';
 import { ExternalApiException } from 'src/common/errors/external-api.error';
-import { WeatherApiWeatherFetch } from 'src/weather/application/constants/weatherapi-weather-fetch.type';
+import { WeatherApiWeatherFetch } from 'src/weather/infrastructure/constants/weatherapi-weather-fetch.type';
 import { ChainableWeatherProvider } from '../../application/interfaces/chainable-weather-provider.abstract';
 import { Weather } from '../../domain/weather.entity';
 import { WeatherApiDtoMapper } from '../mappers/weatherapi.mapper';
